+++ Changelog.md
@@ -1,5 +1,6 @@
 ## 0.0.66
-*INCOMPLETE*
++ Added `beta.CoreMLInferenceSession.exporter` field to support TorchScript and Dynamo ONNX export from PyTorch.
++ Added `beta.OnnxRuntimeInferenceSession.exporter` field to support TorchScript and Dynamo ONNX export from PyTorch.
 
 ## 0.0.65
 + Fixed `Parameter` annotations throwing Pydantic errors when used.